+++ sdk/resourcemanager/servicenetworking/armservicenetworking/client_factory.go
@@ -17,8 +17,7 @@ import (
 // Don't use this type directly, use NewClientFactory instead.
 type ClientFactory struct {
 	subscriptionID string
-	credential     azcore.TokenCredential
-	options        *arm.ClientOptions
+	internal       *arm.Client
 }
 
 // NewClientFactory creates a new instance of ClientFactory with the specified values.
@@ -27,36 +26,51 @@ type ClientFactory struct {
 //   - credential - used to authorize requests. Usually a credential from azidentity.
 //   - options - pass nil to accept the default values.
 func NewClientFactory(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) (*ClientFactory, error) {
-	_, err := arm.NewClient(moduleName, moduleVersion, credential, options)
+	internal, err := arm.NewClient(moduleName, moduleVersion, credential, options)
 	if err != nil {
 		return nil, err
 	}
 	return &ClientFactory{
-		subscriptionID: subscriptionID, credential: credential,
-		options: options.Clone(),
+		subscriptionID: subscriptionID,
+		internal:       internal,
 	}, nil
 }
 
 // NewAssociationsInterfaceClient creates a new instance of AssociationsInterfaceClient.
 func (c *ClientFactory) NewAssociationsInterfaceClient() *AssociationsInterfaceClient {
-	subClient, _ := NewAssociationsInterfaceClient(c.subscriptionID, c.credential, c.options)
-	return subClient
+	return &AssociationsInterfaceClient{
+		subscriptionID: c.subscriptionID,
+		internal:       c.internal,
+	}
 }
 
 // NewFrontendsInterfaceClient creates a new instance of FrontendsInterfaceClient.
 func (c *ClientFactory) NewFrontendsInterfaceClient() *FrontendsInterfaceClient {
-	subClient, _ := NewFrontendsInterfaceClient(c.subscriptionID, c.credential, c.options)
-	return subClient
+	return &FrontendsInterfaceClient{
+		subscriptionID: c.subscriptionID,
+		internal:       c.internal,
+	}
 }
 
 // NewOperationsClient creates a new instance of OperationsClient.
 func (c *ClientFactory) NewOperationsClient() *OperationsClient {
-	subClient, _ := NewOperationsClient(c.credential, c.options)
-	return subClient
+	return &OperationsClient{
+		internal: c.internal,
+	}
+}
+
+// NewSecurityPoliciesInterfaceClient creates a new instance of SecurityPoliciesInterfaceClient.
+func (c *ClientFactory) NewSecurityPoliciesInterfaceClient() *SecurityPoliciesInterfaceClient {
+	return &SecurityPoliciesInterfaceClient{
+		subscriptionID: c.subscriptionID,
+		internal:       c.internal,
+	}
 }
 
 // NewTrafficControllerInterfaceClient creates a new instance of TrafficControllerInterfaceClient.
 func (c *ClientFactory) NewTrafficControllerInterfaceClient() *TrafficControllerInterfaceClient {
-	subClient, _ := NewTrafficControllerInterfaceClient(c.subscriptionID, c.credential, c.options)
-	return subClient
+	return &TrafficControllerInterfaceClient{
+		subscriptionID: c.subscriptionID,
+		internal:       c.internal,
+	}
 }